+++ setup.py
@@ -2,17 +2,16 @@ from setuptools import setup
 
 setup(
     name='gimbleprep',
-    version='0.0.2b3',
+    version='0.0.2b4',
     description='Preprocess fasta, bam and vcf files ready to be used by gimble',
     url='http://github.com/LohseLab/gimbleprep',
     author='Lohse Lab',
     author_email='',
     license='GPLv3',
-    packages=["cli", "lib"],
+    packages=["gimbleprep"],
     package_dir={
         "": ".",
-        "cli": "./cli",
-        "lib": "./lib",
+        "gimbleprep": "./gimbleprep",
     },
     install_requires=[
         "docopt",
@@ -29,7 +28,7 @@ setup(
     ],
     entry_points={
         'console_scripts': [
-            'gimbleprep = cli.interface:main',
+            'gimbleprep = gimbleprep.interface:main',
         ]
     }
 )
